requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.50
pydantic==1.10.7
python-multipart==0.0.6
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
scikit-learn==1.2.2
numpy==1.26.2
